# docs/api-spec.yaml
openapi: "3.0.3"
info:
  title: Crawler.io API
  description: Unified Social Media Data API — one integration, all platforms.
  version: "1.0.0"
  contact:
    email: api@crawler.io
    url: https://docs.crawler.io
servers:
  - url: https://api.crawler.io/v1
    description: Production
  - url: https://sandbox.crawler.io/v1
    description: Sandbox (mock data)
security:
  - BearerAuth: []
...
paths:
...
  /creators/{creator_id}:
    get:
      tags: [Creators]
      summary: Get creator by internal ID
      parameters:
        - name: creator_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^cr_[a-z0-9]+$"
      responses:
        "200":
          description: Creator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /posts/{post_id}:
    get:
      tags: [Posts]
      summary: Get single post by ID
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^po_[a-z0-9]+$"
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"
  /posts/search:
    get:
      tags: [Posts]
      summary: Search posts
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Keyword search in post text
        - name: hashtag
          in: query
          schema:
            type: string
          description: Filter by hashtag (without #)
        - name: platform
          in: query
          schema:
            type: string
          description: Comma-separated platforms
        - name: min_likes
          in: query
          schema:
            type: integer
        - name: min_views
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [image, carousel, video, reel, short, story, live]
        - name: published_after
          in: query
          schema:
            type: string
            format: date
        - name: published_before
          in: query
          schema:
            type: string
            format: date
        - name: sort
          in: query
          schema:
            type: string
            enum: ["-published_at", "-like_count", "-view_count", "relevance"]
            default: "relevance"
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
...
  /tracking/creators:
    post:
      tags: [Tracking]
      summary: Start tracking a creator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, username]
              properties:
                platform:
                  type: string
                  enum: [instagram, tiktok, youtube]
                username:
                  type: string
                frequency:
                  type: string
                  enum: [realtime, frequent, standard]
                  default: standard
                  description: "realtime=30s, frequent=5min, standard=1h"
            example:
              platform: "instagram"
              username: "johndoe"
              frequency: "realtime"
      responses:
        "201":
          description: Tracking started
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  creator_id:
                    type: string
                  status:
                    type: string
                  frequency:
                    type: string
                  created_at:
                    type: string
                    format: date-time
    get:
      tags: [Tracking]
      summary: List tracked creators
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, paused, all]
            default: active
        - name: platform
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Tracked creators list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /tracking/creators/{tracking_id}:
    delete:
      tags: [Tracking]
      summary: Stop tracking a creator
      parameters:
        - name: tracking_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Tracking stopped

  # ─── WEBHOOKS ───────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Pass your key as a Bearer token.
        Production keys: sk_live_*
        Sandbox keys: sk_test_*
  schemas:
    Creator:
      type: object
      properties:
        id:
          type: string
        platform:
          type: string
        platform_id:
          type: string
        username:
          type: string
        display_name:
          type: string
        bio:
          type: string
        profile_image_url:
          type: string
        is_verified:
          type: boolean
        follower_count:
          type: integer
        following_count:
          type: integer
        post_count:
          type: integer
        engagement_rate:
          type: number
        categories:
          type: array
          items:
            type: string
        external_urls:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
        first_seen_at:
          type: string
          format: date-time
        last_updated_at:
          type: string
          format: date-time
        data_freshness:
          type: string
          enum: [live, recent, stale]
    Post:
      type: object
      properties:
        id:
          type: string
        creator_id:
          type: string
        platform:
          type: string
        platform_post_id:
          type: string
        type:
          type: string
        content:
          type: object
          properties:
            text:
              type: string
            hashtags:
              type: array
              items:
                type: string
            mentions:
              type: array
              items:
                type: string
            media:
              type: array
              items:
                $ref: "#/components/schemas/Media"
        metrics:
          $ref: "#/components/schemas/PostMetrics"
        published_at:
          type: string
          format: date-time
        collected_at:
          type: string
          format: date-time
        metrics_updated_at:
          type: string
          format: date-time
    Media:
      type: object
      properties:
        type:
          type: string
          enum: [image, video]
        url:
          type: string
        thumbnail_url:
          type: string
        duration_seconds:
          type: number
        width:
          type: integer
        height:
          type: integer
    PostMetrics:
      type: object
      properties:
        like_count:
          type: integer
        comment_count:
          type: integer
        share_count:
          type: integer
        view_count:
          type: integer
        save_count:
          type: integer
        engagement_rate:
          type: number
...
    Pagination:
      type: object
      properties:
        has_more:
          type: boolean
        next_cursor:
          type: string
        total_count:
          type: integer
...
    PaginatedPosts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        pagination:
          $ref: "#/components/schemas/Pagination"
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            status:
              type: integer
            retry_after:
              type: integer
        request_id:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "not_found"
              message: "Creator not found"
              status: 404
            request_id: "req_abc123"
...